Adjusted code according to mentors commentaries
