+++ Engine/Engine.vcxproj.filters
@@ -18,6 +18,15 @@
     <ClCompile Include="src\main.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="math\sphere.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="src\window.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="src\actor.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="math\ray.h">
@@ -32,7 +41,7 @@
     <ClInclude Include="math\hitable.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\Actor.h">
+    <ClInclude Include="src\actor.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="math\mat4.h">
@@ -41,7 +50,10 @@
     <ClInclude Include="math\vec4.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\Renderer.h">
+    <ClInclude Include="src\renderer.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\window.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
