+++ Engine/math/ray.h
@@ -1,5 +1,4 @@
-#ifndef RAY
-#define RAY
+#pragma once
 
 #include "vec3.h"
 
@@ -10,17 +9,17 @@ class ray
 {
 public:
 	vec3 A;
-	vec3 B;
+	vec3 B; // always normalized
 
 	ray() {}
 	ray(const vec3& a, const vec3& b)
-		: A(a), B(b)
-	{}
+		: A(a)
+	{
+		B = unitVector(b);
+//		B = b;
+	}
 	
 	vec3 origin()    const { return A; }
 	vec3 direction() const { return B; }
 	vec3 pointAtParameter(float t) const { return A + t * B;}
 };
-
-
-#endif
